Added the SercosNetworkBridge batch file
This batch file allows a user to connect to a device (or devices) on an isolated Sercos III network through an MLC's (IndraControl L or XM) internal routing. I.e. it is possible to connect to a Sercos device AND an MLC by plugging a single Ethernet cable into the MLC's engineering  port, or a switch connected to this port. The IP address of the computer running this script, and connecting to these devices must have an IP address assigned to it, which is on the same network as the engineering port of the MLC.